+++ src/controllers/SettingsController.php
@@ -48,7 +48,7 @@ class SettingsController extends Controller
     public function init()
     {
         // Require permission
-        $this->requirePermission('campaign-accessSettings');
+        $this->requirePermission('campaign-settings');
 
         $this->_settings = Campaign::$plugin->getSettings();
     }