+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Unreleased
 ### Added
 - Added preflight sendout screen
+- Added more user group permissions
 ### Changed
 - Changed "Settings" to "Plugin Settings"
 - Changed "userId" to "senderId" in Sendouts table